fix: Moves json seriliazation into DBPersistentManager thread
