+++ core/src/test/java/com/rudderstack/android/sdk/core/EventRepositoryTest.java
@@ -257,120 +257,5 @@ public class EventRepositoryTest {
                         hasEntry("All", true)
                 ));
     }
-
-    @Test
-    public void testGetEventJson() throws JSONException {
-        EventRepository repo = new EventRepository();
-        HashMap jsonData = new HashMap<String, Object>();
-        jsonData.put("pinId", 2351636);
-        jsonData.put("userLocationLongitude", 12.516869940636775);
-        jsonData.put("userLocationLatitude", 55.661663449562205);
-        jsonData.put("userDirection", 12.311176);
-        jsonData.put("userSpeed", 13.888889);
-        jsonData.put("locationHorizontalAccuracy", 3.813398);
-        jsonData.put("locationVerticalAccuracy", 0.0);
-        jsonData.put("speedAccuracy", 0.0);
-        jsonData.put("directionAccuracy", 0.0);
-        JSONArray locationsBefore = new JSONArray();
-        locationsBefore.put(new JSONObject().put("latitude", 55.66132122924984).put("longitude", 12.51671169784383));
-        locationsBefore.put(new JSONObject().put("latitude", 55.661428115890374).put("longitude", 12.51677390468785));
-        locationsBefore.put(new JSONObject().put("latitude", 55.661663449562205).put("longitude", 12.516869940636775));
-        jsonData.put("locationsBefore", locationsBefore);
-        JSONArray locationsAfter = new JSONArray();
-        locationsAfter.put(new JSONObject().put("latitude", 55.66190443423447).put("longitude", 12.516850445696493));
-        locationsAfter.put(new JSONObject().put("latitude", 55.66214021085126).put("longitude", 12.516731010346978));
-        locationsAfter.put(new JSONObject().put("latitude", 55.66237844832189).put("longitude", 12.516618424859859));
-        jsonData.put("locationsAfter", locationsAfter);
-        JSONArray speedsBefore = new JSONArray();
-        speedsBefore.put(new JSONObject().put("speed", 0));
-        speedsBefore.put(new JSONObject().put("speed", 13.888889));
-        speedsBefore.put(new JSONObject().put("speed", 13.888889));
-        jsonData.put("speedsBefore", speedsBefore);
-        JSONArray speedsAfter = new JSONArray();
-        speedsAfter.put(new JSONObject().put("speed", 13.888889));
-        speedsAfter.put(new JSONObject().put("speed", 13.888889));
-        jsonData.put("speedsAfter", speedsAfter);
-        jsonData.put("pinType", 1);
-        jsonData.put("pinSubtype", 11);
-        jsonData.put("pinDirection", 0.0);
-        jsonData.put("pinLocationLongitude", 12.507275);
-        jsonData.put("pinLocationLatitude", 55.672436);
-        jsonData.put("coDriverVersion", -1.0);
-        RudderMessage message = new RudderMessageBuilder().setEventName("TestEvent").setProperty(jsonData).build();
-        String expectedJsonString = "{\n" +
-                "  \"messageId\": \"" + message.getMessageId() + "\",\n" +
-                "  \"channel\": \"mobile\",\n" +
-                "  \"context\": {},\n" +
-                "  \"originalTimestamp\": \"2022-03-14T06:46:41.365Z\",\n" +
-                "  \"event\": \"TestEvent\",\n" +
-                "  \"properties\": {\n" +
-                "    \"pinLocationLongitude\": 12.507275,\n" +
-                "    \"userDirection\": 12.311176,\n" +
-                "    \"locationHorizontalAccuracy\": 3.813398,\n" +
-                "    \"pinId\": 2351636,\n" +
-                "    \"pinDirection\": 0.0,\n" +
-                "    \"coDriverVersion\": -1.0,\n" +
-                "    \"userLocationLatitude\": 55.661663449562205,\n" +
-                "    \"userSpeed\": 13.888889,\n" +
-                "    \"locationsBefore\": [\n" +
-                "      {\n" +
-                "        \"latitude\": 55.66132122924984,\n" +
-                "        \"longitude\": 12.51671169784383\n" +
-                "      },\n" +
-                "      {\n" +
-                "        \"latitude\": 55.661428115890374,\n" +
-                "        \"longitude\": 12.51677390468785\n" +
-                "      },\n" +
-                "      {\n" +
-                "        \"latitude\": 55.661663449562205,\n" +
-                "        \"longitude\": 12.516869940636775\n" +
-                "      }\n" +
-                "    ],\n" +
-                "    \"speedAccuracy\": 0.0,\n" +
-                "    \"locationVerticalAccuracy\": 0.0,\n" +
-                "    \"pinType\": 1,\n" +
-                "    \"speedsAfter\": [\n" +
-                "      {\n" +
-                "        \"speed\": 13.888889\n" +
-                "      },\n" +
-                "      {\n" +
-                "        \"speed\": 13.888889\n" +
-                "      }\n" +
-                "    ],\n" +
-                "    \"pinSubtype\": 11,\n" +
-                "    \"locationsAfter\": [\n" +
-                "      {\n" +
-                "        \"latitude\": 55.66190443423447,\n" +
-                "        \"longitude\": 12.516850445696493\n" +
-                "      },\n" +
-                "      {\n" +
-                "        \"latitude\": 55.66214021085126,\n" +
-                "        \"longitude\": 12.516731010346978\n" +
-                "      },\n" +
-                "      {\n" +
-                "        \"latitude\": 55.66237844832189,\n" +
-                "        \"longitude\": 12.516618424859859\n" +
-                "      }\n" +
-                "    ],\n" +
-                "    \"speedsBefore\": [\n" +
-                "      {\n" +
-                "        \"speed\": 0\n" +
-                "      },\n" +
-                "      {\n" +
-                "        \"speed\": 13.888889\n" +
-                "      },\n" +
-                "      {\n" +
-                "        \"speed\": 13.888889\n" +
-                "      }\n" +
-                "    ],\n" +
-                "    \"pinLocationLatitude\": 55.672436,\n" +
-                "    \"userLocationLongitude\": 12.516869940636775,\n" +
-                "    \"directionAccuracy\": 0.0\n" +
-                "  },\n" +
-                "  \"integrations\": {}\n" +
-                "}";
-        String outputJsonString = repo.getEventJsonString(message);
-        assertThat("JSONObjects and JSONArray are serialized perfectly", outputJsonString, is(expectedJsonString.replace("\n", "").replace(" ", "")));
-    }
 }
